+++ CHANGELOG.md
@@ -14,6 +14,7 @@ NOTE: Worlds that used Refined Storage 1.5.x are fully compatible with Refined S
 - Added new storage disk system where the storage disk data (items, fluids) are stored off the disk itself, in another file (refinedstorage_disks.dat). The disk itself only stores its ID (raoulvdberge)
 - Added /createdisk command which creates a disk based on the disk ID. Turn on advanced tooltips to see the disk ID on a disk item (raoulvdberge)
 - Added config option to configure energy capacity of Refined Storage items (raoulvdberge)
+- Added config option to change Reader / Writer channel energy capacity (raoulvdberge)
 - Changed fluid storage progression to be 64k - 256k - 1024k - 4096k (raoulvdberge)
 - You can no longer put a Filter in filter slots to gain additional filter slots (raoulvdberge)
 - You can now re-insert Processing Patterns in the Pattern Grid and have the inputs and outputs be completed (raoulvdberge)